+++ src/data/urls.ts
@@ -1,0 +1,9 @@
+export const DEFAULT_SYMBOLS_URL =
+  "https://raw.githubusercontent.com/ttqteo/vnstock-js/master/data/symbols.json";
+
+export const DEFAULT_HOLIDAYS_URL =
+  "https://raw.githubusercontent.com/ttqteo/vnstock-js/master/data/holidays.json";
+
+export const DEFAULT_TTL_MS = 24 * 60 * 60 * 1000;
+
+export const DEFAULT_FETCH_TIMEOUT_MS = 10000;